+++ flutter/shell/platform/tizen/flutter_tizen_engine.h
@@ -21,6 +21,7 @@
 #include "flutter/shell/platform/tizen/channels/platform_view_channel.h"
 #include "flutter/shell/platform/tizen/channels/settings_channel.h"
 #include "flutter/shell/platform/tizen/flutter_project_bundle.h"
+#include "flutter/shell/platform/tizen/flutter_tizen_display_monitor.h"
 #include "flutter/shell/platform/tizen/flutter_tizen_texture_registrar.h"
 #include "flutter/shell/platform/tizen/public/flutter_tizen.h"
 #include "flutter/shell/platform/tizen/tizen_event_loop.h"
@@ -48,7 +49,6 @@ namespace flutter {
 constexpr FlutterViewId kImplicitViewId = 0;
 
 class FlutterTizenView;
-class FlutterTizenDisplayMonitor;
 
 // Manages state associated with the underlying FlutterEngine.
 class FlutterTizenEngine {
