+++ flutter/shell/platform/tizen/flutter_tizen_view.cc
@@ -6,6 +6,7 @@
 #include "flutter_tizen_view.h"
 
 #include "flutter/shell/platform/tizen/logger.h"
+#include "flutter/shell/platform/tizen/system_utils.h"
 #include "flutter/shell/platform/tizen/tizen_view.h"
 #ifdef NUI_SUPPORT
 #include "flutter/shell/platform/tizen/tizen_view_nui.h"
@@ -13,15 +14,7 @@
 #include "flutter/shell/platform/tizen/tizen_renderer_egl.h"
 #include "flutter/shell/platform/tizen/tizen_window.h"
 
-namespace flutter {
-
-#if defined(MOBILE_PROFILE)
-constexpr double kProfileFactor = 0.7;
-#elif defined(TV_PROFILE)
-constexpr double kProfileFactor = 2.0;
-#else
-constexpr double kProfileFactor = 1.0;
-#endif
+namespace {
 
 constexpr char kSysMenuKey[] = "XF86SysMenu";
 constexpr char kBackKey[] = "XF86Back";
@@ -45,18 +38,9 @@ const std::vector<std::string> kBindableSystemKeys = {
 // (ui/events/x/events_x_utils.cc).
 constexpr int32_t kScrollOffsetMultiplier = 53;
 
-double ComputePixelRatio(int32_t screen_dpi) {
-  // The scale factor is computed based on the display DPI and the current
-  // profile. A fixed DPI value (72) is used on TVs. See:
-  // https://docs.tizen.org/application/native/guides/ui/efl/multiple-screens
-#ifdef TV_PROFILE
-  double dpi = 72.0;
-#else
-  double dpi = static_cast<double>(screen_dpi);
-#endif
-  double scale_factor = dpi / 90.0 * kProfileFactor;
-  return std::max(scale_factor, 1.0);
-}
+}  // namespace
+
+namespace flutter {
 
 FlutterTizenView::FlutterTizenView(FlutterViewId view_id,
                                    std::unique_ptr<TizenViewBase> tizen_view,
