+++ flutter/shell/platform/tizen/system_utils.h
@@ -10,6 +10,14 @@
 
 namespace flutter {
 
+#if defined(MOBILE_PROFILE)
+constexpr double kProfileFactor = 0.7;
+#elif defined(TV_PROFILE)
+constexpr double kProfileFactor = 2.0;
+#else
+constexpr double kProfileFactor = 1.0;
+#endif
+
 // Components of a system language/locale.
 struct LanguageInfo {
   std::string language;
@@ -17,6 +25,10 @@ struct LanguageInfo {
   std::string script;
   std::string variant;
 };
+
+// Returns the scale factor calculated based on the display DPI and the current
+// profile.
+double ComputePixelRatio(int32_t screen_dpi);
 
 // Returns the list of user-preferred locales, in preference order,
 // parsed into LanguageInfo structures.
